BAH-4721 | Refactor. Move DURATION_UNIT_OPTIONS Constant Into MedicationRequest Constant
