+++ apps/clinical/src/constants/medications.ts
@@ -1,16 +1,2 @@
-import { DurationUnitOption } from '../models/medication';
-
 // Duration of the boundsPeriod validity window for STAT (immediate) orders
 export const STAT_ORDER_VALIDITY_MS = 5 * 60 * 1000; // 5 Minutes
-
-export const DURATION_UNIT_OPTIONS: DurationUnitOption[] = [
-  {
-    code: 'min',
-    display: 'DURATION_UNIT_MINUTES',
-    daysMultiplier: 1 / (24 * 60),
-  },
-  { code: 'h', display: 'DURATION_UNIT_HOURS', daysMultiplier: 1 / 24 },
-  { code: 'd', display: 'DURATION_UNIT_DAYS', daysMultiplier: 1 },
-  { code: 'wk', display: 'DURATION_UNIT_WEEKS', daysMultiplier: 7 },
-  { code: 'mo', display: 'DURATION_UNIT_MONTHS', daysMultiplier: 30 },
-];
